+++ CHANGELOG.md
@@ -31,6 +31,7 @@ Changelog for NeoFS Node
 - Object left in the storage after unsuccessful meta PUT (#3744)
 - Object left in some shards on fatal broadcast failure (#3744)
 - Deadlock on SN exit in rare GC cases (#3744)
+- SN no longer counts removed parent objects as physical ones in metrics (#3717)
 
 ### Changed
 - Optimized locking info in metabase (#3672)